chore: remove SkipListner and ChunkListener from LogStepListener
